// frontend/person-blog-app/packages/blog/src/shared/api/generated/models/createPostModelViewModel.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * Gateway.API
 * OpenAPI spec version: 1.0
 */
import type { CategoryModel } from './categoryModel';
import type { PostVisibilitySelectItem } from './postVisibilitySelectItem';
import type { SubscriptionLevelModel } from './subscriptionLevelModel';

export interface CreatePostModelViewModel {
  /** @nullable */
  subscriptionLevels?: SubscriptionLevelModel[] | null;
  /** @nullable */
  visibility?: PostVisibilitySelectItem[] | null;
  /** @nullable */
  categoryList?: CategoryModel[] | null;
}